CH4/demo1/pyproject.toml
[tool.poetry]
name = "demo1"
version = "0.1.0"
description = ""
authors = ["Ian Chen <xxx@xxx.xxx>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain-openai = "0.2.0"
langchain = "0.3.0"
langchain-core = "0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
